fix: update ViewersCoreLink to also work out of MDAWidget
