+++ src/pymmcore_gui/core_link/__init__.py
@@ -1,4 +1,4 @@
-# from ._viewers_core_link import ViewersCoreLink
-from ._viewers_core_link_v2 import ViewersCoreLink
+from ._viewers_core_link import ViewersCoreLink
+# from ._viewers_core_link_v2 import ViewersCoreLink
 
 __all__ = ["ViewersCoreLink"]
